+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## _(in-progress)_
 * New: Add TODO.md and move existing TODO list from top of main plugin file into it
+* Change: Remove long-deprecated `inject_query_posts()`
 * Change: Note compatibility through WP 5.5+
 
 ## 2.2.9 _(2020-05-01)_